src/requirements.txt
opencv-python
imgbeddings
psycopg2-binary
click
